🪚Refactor Gemini/Vertex handlers
Combine common functionality between GeminiHandler and VertexAIHandler into a common GeminiHandlerBase class. GeminiHandler subclasses that and adds a concrete Gemini provider using the production Gemini URLs and an API key. When creating a Vertex AI provider, two possible handler classes are used. If the model name begins with “gemini”, a VertexGeminiHandler is used. If the model begins with “claude” a separate VertexClaudeHandler is used. Those two handlers reuse existing Gemini and Claude response types.
